separate UDP and TCP queries for DNS Server
